Added filtering pagination to sources table.
